moved all strategy classes to strategies/
