+++ app.py
@@ -1474,338 +1474,6 @@ class Bot:
         return False
 
 
-class BuyMoonSellRecoveryStrategy(Bot):
-    """Base Strategy Class"""
-
-    def buy_strategy(self, coin: Coin) -> bool:
-        """bot buy strategy"""
-        if float(coin.price) > percent(coin.buy_at_percentage, coin.last):
-            self.buy_coin(coin)
-            self.log_debug_coin(coin)
-            return True
-        return False
-
-
-class BuyOnGrowthTrendAfterDropStrategy(Bot):
-    """Buy Strategy
-
-    Wait for a coin to drop below BUY_AT_PERCENTAGE and then
-    monitor its growth trend over a certain period, where each slice of
-    that period must grow by at least n% over the previous slice.
-    As soon that happens buy this coin.
-    """
-
-    def buy_strategy(self, coin: Coin) -> bool:
-        """bot buy strategy"""
-
-        unit = str(coin.klines_trend_period[-1:]).lower()
-        klines_trend_period = int(''.join(coin.klines_trend_period[:-1]))
-        last_period = list(coin.averages[unit])[-klines_trend_period:]
-
-        if len(last_period) < klines_trend_period:
-            return False
-
-        # has the price gone down by x% on a coin we don't own?
-        if (
-            (float(coin.price) < percent(coin.buy_at_percentage, coin.max))
-            and coin.status == ""
-            and not coin.naughty
-        ):
-            coin.dip = coin.price
-            coin.status = "TARGET_DIP"
-            logging.info(
-                f"{c_from_timestamp(coin.date)}: {coin.symbol} [{coin.status}] "
-                + f"-> [TARGET_DIP] ({coin.price})"
-            )
-
-        if coin.status != "TARGET_DIP":
-            return False
-
-        # if the price keeps going down, skip it
-        last_period_slice = last_period[0][1]
-        for _, n in last_period[1:]:
-            if (
-                percent(
-                    100 + coin.klines_slice_percentage_change,
-                    last_period_slice,
-                )
-                > n
-            ):
-                return False
-            last_period_slice = n
-        self.buy_coin(coin)
-        return True
-
-class BuyOnRecoveryAfterDropDuringGrowthTrendStrategy(Bot):
-    """Buy Strategy
-
-    """
-
-    def buy_strategy(self, coin: Coin) -> bool:
-        """bot buy strategy"""
-
-        unit = str(coin.klines_trend_period[-1:]).lower()
-        klines_trend_period = int(coin.klines_trend_period[:-1])
-
-        last_period = list(coin.averages[unit])[-klines_trend_period:]
-
-        if len(last_period) < klines_trend_period:
-            return False
-
-        last_period_slice = last_period[0][1]
-        # if the price keeps going down, skip it
-        for _, n in last_period[1:]:
-            if (
-                percent(
-                    100 + coin.klines_slice_percentage_change,
-                    last_period_slice,
-                )
-                > n
-            ):
-                return False
-            last_period_slice = n
-
-        # has the price gone down by x% on a coin we don't own?
-        if (
-            (float(coin.price) < percent(coin.buy_at_percentage, coin.max))
-            and coin.status == ""
-            and not coin.naughty
-        ):
-            coin.dip = coin.price
-            logging.info(
-                f"{c_from_timestamp(coin.date)}: {coin.symbol} [{coin.status}] "
-                + f"-> [TARGET_DIP] ({coin.price})"
-            )
-            coin.status = "TARGET_DIP"
-            return False
-
-        if coin.status != "TARGET_DIP":
-            return False
-
-        # do some gimmicks, and don't buy the coin straight away
-        # but only buy it when the price is now higher than the last
-        # price recorded. This way we ensure that we got the dip
-        self.log_debug_coin(coin)
-        if float(coin.price) > float(coin.last):
-            if float(coin.price) > percent(
-                float(coin.trail_recovery_percentage), coin.dip
-            ):
-                self.buy_coin(coin)
-                return True
-        return False
-
-class BuyDropSellRecoveryStrategy(Bot):
-    """Base Strategy Class"""
-
-    def buy_strategy(self, coin: Coin) -> bool:
-        """bot buy strategy"""
-
-        # has the price gone down by x% on a coin we don't own?
-        if (
-            (float(coin.price) < percent(coin.buy_at_percentage, coin.max))
-            and coin.status == ""
-            and not coin.naughty
-        ):
-            coin.dip = coin.price
-            logging.info(
-                f"{c_from_timestamp(coin.date)}: {coin.symbol} [{coin.status}] "
-                + f"-> [TARGET_DIP] ({coin.price})"
-            )
-            coin.status = "TARGET_DIP"
-
-        if coin.status != "TARGET_DIP":
-            return False
-
-        # do some gimmicks, and don't buy the coin straight away
-        # but only buy it when the price is now higher than the last
-        # price recorded. This way we ensure that we got the dip
-        self.log_debug_coin(coin)
-        if float(coin.price) > float(coin.last):
-            if float(coin.price) > percent(
-                float(coin.trail_recovery_percentage), coin.dip
-            ):
-                self.buy_coin(coin)
-                return True
-        return False
-
-class BuyDropSellRecoveryStrategyWhenBTCisUp(Bot):
-    """Base Strategy Class"""
-
-    def buy_strategy(self, coin: Coin) -> bool:
-        """bot buy strategy"""
-
-        # with this strategy we never buy BTC
-        if coin.symbol == "BTCUSDT":
-            return False
-
-        if 'BTCUSDT' not in self.coins:
-            return False
-
-        unit = str(self.coins['BTCUSDT'].klines_trend_period[-1:]).lower()
-        klines_trend_period = int(
-            ''.join(self.coins['BTCUSDT'].klines_trend_period[:-1])
-        )
-
-        last_period = list(
-            self.coins['BTCUSDT'].averages[unit]
-        )[-klines_trend_period:]
-
-        if len(last_period) < klines_trend_period:
-            return False
-
-        last_period_slice = last_period[0][1]
-        for _, n in last_period[1:]:
-            if (
-                percent(
-                    100 + float(
-                        self.coins['BTCUSDT'].klines_slice_percentage_change
-                    ),
-                    last_period_slice,
-                )
-                > n
-            ):
-                return False
-            last_period_slice = n
-
-        # has the price gone down by x% on a coin we don't own?
-        if (
-            (float(coin.price) < percent(coin.buy_at_percentage, coin.max))
-            and coin.status == ""
-            and not coin.naughty
-        ):
-            coin.dip = coin.price
-            logging.info(
-                f"{c_from_timestamp(coin.date)}: {coin.symbol} [{coin.status}] "
-                + f"-> [TARGET_DIP] ({coin.price})"
-            )
-            coin.status = "TARGET_DIP"
-
-        if coin.status != "TARGET_DIP":
-            return False
-
-        # do some gimmicks, and don't buy the coin straight away
-        # but only buy it when the price is now higher than the last
-        # price recorded. This way we ensure that we got the dip
-        self.log_debug_coin(coin)
-        if float(coin.price) < float(coin.last):
-            if float(coin.price) > percent(
-                float(coin.trail_recovery_percentage), coin.dip
-            ):
-                self.buy_coin(coin)
-                return True
-        return False
-
-class BuyDropSellRecoveryStrategyWhenBTCisDown(Bot):
-    """Base Strategy Class"""
-
-    def buy_strategy(self, coin: Coin) -> bool:
-        """bot buy strategy"""
-
-        # with this strategy we never buy BTC
-        if coin.symbol == "BTCUSDT":
-            return False
-
-        if 'BTCUSDT' not in self.coins:
-            return False
-
-        unit = str(self.coins['BTCUSDT'].klines_trend_period[-1:]).lower()
-        klines_trend_period = int(self.coins['BTCUSDT'].klines_trend_period[:-1])
-
-        last_period = list(
-            self.coins['BTCUSDT'].averages[unit]
-        )[-klines_trend_period:]
-
-        if len(last_period) < klines_trend_period:
-            return False
-
-        last_period_slice = last_period[0][1]
-        for _, n in last_period[1:]:
-            if (
-                percent(
-                    100 + (
-                        -abs(float(
-                            self.coins['BTCUSDT'].klines_slice_percentage_change
-                        ))
-                    ),
-                    last_period_slice,
-                ) < n
-            ):
-                return False
-            last_period_slice = n
-
-
-        # has the price gone down by x% on a coin we don't own?
-        if (
-            (float(coin.price) < percent(coin.buy_at_percentage, coin.max))
-            and coin.status == ""
-            and not coin.naughty
-        ):
-            coin.dip = coin.price
-            logging.info(
-                f"{c_from_timestamp(coin.date)}: {coin.symbol} [{coin.status}] "
-                + f"-> [TARGET_DIP] ({coin.price})"
-            )
-            coin.status = "TARGET_DIP"
-
-        if coin.status != "TARGET_DIP":
-            return False
-
-        # do some gimmicks, and don't buy the coin straight away
-        # but only buy it when the price is now higher than the last
-        # price recorded. This way we ensure that we got the dip
-        self.log_debug_coin(coin)
-        if float(coin.price) > float(coin.last):
-            if float(coin.price) > percent(
-                float(coin.trail_recovery_percentage), coin.dip
-            ):
-                self.buy_coin(coin)
-                return True
-        return False
-
-class BuyOnRecoveryAfterDropFromAverageStrategy(Bot):
-    """Base Strategy Class"""
-
-    def buy_strategy(self, coin: Coin) -> bool:
-        """bot buy strategy"""
-
-        unit = str(coin.klines_trend_period[-1:]).lower()
-        klines_trend_period = int(''.join(coin.klines_trend_period[:-1]))
-
-        last_period = list(coin.averages[unit])[-klines_trend_period:]
-
-        if len(last_period) < klines_trend_period:
-            return False
-
-        average = mean([v for d,v in last_period])
-        # has the price gone down by x% on a coin we don't own?
-        if (
-            (float(coin.price) < percent(coin.buy_at_percentage, average))
-            and coin.status == ""
-            and not coin.naughty
-        ):
-            coin.dip = coin.price
-            logging.info(
-                f"{c_from_timestamp(coin.date)}: {coin.symbol} [{coin.status}] "
-                + f"-> [TARGET_DIP] ({coin.price})"
-            )
-            coin.status = "TARGET_DIP"
-
-        if coin.status != "TARGET_DIP":
-            return False
-
-        # do some gimmicks, and don't buy the coin straight away
-        # but only buy it when the price is now higher than the last
-        # price recorded. This way we ensure that we got the dip
-        self.log_debug_coin(coin)
-        if float(coin.price) > float(coin.last):
-            if float(coin.price) > percent(
-                float(coin.trail_recovery_percentage), coin.dip
-            ):
-                self.buy_coin(coin)
-                return True
-        return False
-
-
 if __name__ == "__main__":
     try:
         parser = argparse.ArgumentParser()
@@ -1829,40 +1497,6 @@ if __name__ == "__main__":
         else:
             client = Client(secrets["ACCESS_KEY"], secrets["SECRET_KEY"])
 
-        if cfg["STRATEGY"] == "BuyMoonSellRecoveryStrategy":
-            bot = BuyMoonSellRecoveryStrategy(
-                client, args.config, cfg
-            )  # type: ignore
-
-        elif cfg["STRATEGY"] == "BuyOnGrowthTrendAfterDropStrategy":
-            bot = BuyOnGrowthTrendAfterDropStrategy(
-                client, args.config, cfg
-            )  # type: ignore
-
-        elif cfg["STRATEGY"] == "BuyDropSellRecoveryStrategy":
-            bot = BuyDropSellRecoveryStrategy(
-                client, args.config, cfg
-            )  # type: ignore
-
-        elif cfg["STRATEGY"] == "BuyDropSellRecoveryStrategyWhenBTCisUp":
-            bot = BuyDropSellRecoveryStrategyWhenBTCisUp(
-                client, args.config, cfg
-            )  # type: ignore
-
-        elif cfg["STRATEGY"] == "BuyDropSellRecoveryStrategyWhenBTCisDown":
-            bot = BuyDropSellRecoveryStrategyWhenBTCisDown(
-                client, args.config, cfg
-            )  # type: ignore
-
-        elif cfg["STRATEGY"] == "BuyOnRecoveryAfterDropDuringGrowthTrendStrategy":
-            bot = BuyOnRecoveryAfterDropDuringGrowthTrendStrategy(
-                client, args.config, cfg
-            )  # type: ignore
-
-        elif cfg["STRATEGY"] == "BuyOnRecoveryAfterDropFromAverageStrategy":
-            bot = BuyOnRecoveryAfterDropFromAverageStrategy(
-                client, args.config, cfg
-            )  # type: ignore
 
 
         logging.info(
